requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
pyoembed==0.1.2
python-oembed==0.2.4
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
Werkzeug==2.2.2
